+++ src/Network/Wai/Predicate/MediaType.hs
@@ -9,7 +9,8 @@ module Network.Wai.Predicate.MediaType where
 
 import Data.ByteString (ByteString)
 import Data.ByteString.Char8 (pack)
-import GHC.TypeLits
+import Data.Singletons
+import Data.Singletons.TypeLits
 
 data Media (t :: Symbol) (s :: Symbol) = Media
     { rawType      :: !ByteString